refactor: roundVoteIsFinished 네이밍을 voteIsFinished로 수정 #261
