+++ frontend/src/mocks/data/voteIsFinished.json
