+++ frontend/src/mocks/handlers/balanceContentHandler.ts
@@ -2,7 +2,7 @@ import { http, HttpResponse } from 'msw';
 
 import BALANCE_CONTENT from '../data/balanceContent.json';
 import MY_GAME_STATUS from '../data/myGameStatus.json';
-import ROUND_VOTE_IS_FINISHED from '../data/roundVoteIsFinished.json';
+import VOTE_IS_FINISHED from '../data/voteIsFinished.json';
 
 import { POLLING_DELAY } from '@/constants/config';
 import { MOCK_API_URL } from '@/constants/url';
@@ -12,15 +12,15 @@ const fetchBalanceContentHandler = () => {
   return HttpResponse.json<BalanceContent>(BALANCE_CONTENT);
 };
 
-const fetchIsFinishedHandler = () => {
+const fetchVoteIsFinishedHandler = () => {
   setTimeout(() => {
-    ROUND_VOTE_IS_FINISHED.isFinished = true;
+    VOTE_IS_FINISHED.isFinished = true;
   }, 10 * POLLING_DELAY);
   setTimeout(() => {
-    ROUND_VOTE_IS_FINISHED.isFinished = false;
+    VOTE_IS_FINISHED.isFinished = false;
   }, 12 * POLLING_DELAY);
 
-  return HttpResponse.json(ROUND_VOTE_IS_FINISHED);
+  return HttpResponse.json(VOTE_IS_FINISHED);
 };
 
 const getMyGameStatus = ({ request }: { request: Request }) => {
@@ -29,6 +29,6 @@ const getMyGameStatus = ({ request }: { request: Request }) => {
 
 export const contentHandler = [
   http.get(MOCK_API_URL.balanceContent, fetchBalanceContentHandler),
-  http.get(MOCK_API_URL.roundVoteIsFinished, fetchIsFinishedHandler),
+  http.get(MOCK_API_URL.voteIsFinished, fetchVoteIsFinishedHandler),
   http.get(MOCK_API_URL.myGameStatus, getMyGameStatus),
 ];
