+++ frontend/src/mocks/handlers/voteHandler.ts
@@ -3,15 +3,15 @@ import { http, HttpResponse } from 'msw';
 import BALANCE_CONTENT from '../data/balanceContent.json';
 import FINAL_RESULT from '../data/finalResult.json';
 import MY_GAME_STATUS from '../data/myGameStatus.json';
-import ROUND_VOTE_IS_FINISHED from '../data/roundVoteIsFinished.json';
 import VOTE_RESULT from '../data/roundVoteResult.json';
+import VOTE_IS_FINISHED from '../data/voteIsFinished.json';
 
 import { MOCK_API_URL } from '@/constants/url';
 import { RoundVoteResult } from '@/types/roundVoteResult';
 
 const voteBalanceContentHandler = async ({ request }: { request: Request }) => {
   const body = await request.json();
-  ROUND_VOTE_IS_FINISHED.isFinished = false;
+  VOTE_IS_FINISHED.isFinished = false;
   return HttpResponse.json(body, { status: 201 });
 };
 
@@ -22,7 +22,7 @@ const fetchVoteResultHandler = async () => {
 const goToNextRoundHandler = () => {
   BALANCE_CONTENT.currentRound += 1;
   MY_GAME_STATUS.isRoundFinished = true;
-  ROUND_VOTE_IS_FINISHED.isFinished = false;
+  VOTE_IS_FINISHED.isFinished = false;
   return HttpResponse.json({ state: 204 });
 };
